Organize stuff and resolve burnable bug
