+++ src/main/java/com/sbancuz/plannh/Compat.java
@@ -7,7 +7,7 @@ import com.sbancuz.plannh.data.PropertyProvider;
 import com.sbancuz.plannh.data.provider.BotaniaProvider;
 import com.sbancuz.plannh.data.provider.EnderIOProvider;
 import com.sbancuz.plannh.data.provider.ForestryProvider;
-import com.sbancuz.plannh.data.provider.GTProvider;
+import com.sbancuz.plannh.data.provider.gregtech.GTProvider;
 import com.sbancuz.plannh.data.provider.ThaumcraftProvider;
 
 import cpw.mods.fml.common.Loader;
