+++ src/main/java/com/sbancuz/plannh/data/flowchart/Note.java
@@ -6,11 +6,9 @@ import javax.annotation.Nonnull;
 
 public class Note {
 
-    @Nonnull
     public final UUID id;
     public int x;
     public int y;
-    @Nonnull
     public String text = "";
 
     public Note(final UUID id, final int x, final int y) {
